fix(dtu): static stub manifest (no gh) written before vite for gitea e2e
Follow-up to #399. That PR fixed the L3 app-boot 500 by writing a stub
src/generated/repo-manifest.json from twins/gitea/global-setup.ts. Two
refinements here:

1. The globalSetup placement is racy: Playwright starts the webServers
   concurrently with (in CI, before) globalSetup, so Vite pre-transforms
   src/engine/local-loader.ts and logs a transient "Failed to resolve import
   ../generated/repo-manifest.json" before the stub lands. Specs still pass
   (Vite self-heals once the file exists), but the error string remains in the
   dev-server log.
2. An earlier iteration ran scripts/generate-manifest.js, which shells out to
   `gh` on every L3 boot — it degrades to empty without GH_TOKEN but adds a
   network/CLI failure mode to the nightly lane.

Move the stub-write into a committed, dependency-free script
twins/gitea/ensure-manifest-stub.mjs and chain it ahead of `vite` in the
playwright.gitea.config.ts webServer command:
- Race-free: the file exists before Vite's first transform (no transient error).
- Offline/deterministic: a static stub, no generate-manifest.js, no `gh`.
- Idempotent: skips when the file already exists, so a real local-mode manifest
  (e.g. from `prebuild`) is never clobbered.

The stub carries only the required RepoManifest fields (empty). L3 runs the app
in REMOTE mode (no VITE_KB_LOCAL), so the manifest is never read at runtime — it
exists solely to satisfy Vite's import resolver. The path is gitignored, so it
never enters the seed's git add -A snapshot. Reverts the now-redundant
stub-write from global-setup.ts.

Refs #394
